API Reference

Finding Tikhonov-Fenichel Parameter Values

TikhonovFenichelReductions.ReductionProblem — Type

mutable struct ReductionProblem

Type that defines a Tikhonov-Fenichel reduction problem, i.e. a polynomial ODE system, for which all slow-fast separations of rates yielding a reduction onto dimension s are considered.

Fields

f::Vector{QQMPolyRingElem}: RHS of ODE system as a vector of polynomials

x::Vector{QQMPolyRingElem}: Vector of dynamic variables

p::Vector{QQMPolyRingElem}: Vector of all parameters

s::Int: Dimension of reduced system

p_sf::Vector{QQMPolyRingElem}: Vector of parameters to be considered slow or fast (all others are considered fixed)

idx_slow_fast::Vector{Bool}: Boolean index, such that p_sf=p[idx_slow_fast]

J::AbstractAlgebra.Generic.MatSpaceElem{QQMPolyRingElem}: Jacobian of f

_f::Function: RHS of ODE system as a Julia function with arguments x and p

The type QQMPolyRingElem is used in Oscar.jl to represent elements of a polynomial ring; here this is ℝ[x,p].

TikhonovFenichelReductions.ReductionProblem — Method

ReductionProblem(
    f::Function,
    x::Vector{String},
    p::Vector{String},
    s::Int64;
    idx_slow_fast
) -> ReductionProblem

Constructor for ReductionProblem Type.

Arguments

f(x,p)::Function: Julia function defining the RHS of system

x::Vector{String}: Vector of dynamic variables

p::Vector{String}: Vector of all parameters

s::Int: Dimension of reduced system

idx_slow_fast::Vector{Bool}: Boolean index for all rates that are either small or large (all others are considered fixed)

Description

This function is used to set up the problem of finding Tikhonov-Fenichel Parameter Values for dimension s. The names of all variables and parameters in the system are parsed to appropriate types in Oscar.jl, so that the necessary conditions for the existence of such a reduction can be evaluated.

See also: tfpv_candidates

TikhonovFenichelReductions.system_components — Method

system_components(
    problem::ReductionProblem
) -> Vector{Nemo.QQMPolyRingElem}

Convenience function to get components of ODE system from instance of ReductionProblem.

See also: ReductionProblem, system_parameters

TikhonovFenichelReductions.system_parameters — Method

system_parameters(
    problem::ReductionProblem
) -> Vector{Nemo.QQMPolyRingElem}

Convenience function to get parameters of ODE system from instance of ReductionProblem.

See also: ReductionProblem, system_components

TikhonovFenichelReductions.tfpv_candidates — Method

tfpv_candidates(
    problem::ReductionProblem
) -> Tuple{Vector{Vector{Bool}}, Vector, Vector{Vector{Int64}}, Vector{Vector{Bool}}}

Find all slow-fast separations π⁺ that are TFPVs by using the necessary conditions

the affine variety V(f⁰) contains an irreducible component Y of dimension s

the s-th coefficient of the characteristic polynomial of D₁f(x,π⁺) is non-zero for x∈Y

Description

The irreducible components are obtained by computing a minimal primary decomposition. The Jacobian at a point in an irreducible component Y is constructed symbolically by computing normal forms with respect to a Gröbner basis G, s.t. 𝑉(G)=Y.

To obtain all general TFPVs and not just slow-fast separations, one can use the function tfpv_groebner_basis, which relies on computing a Gröbner basis with an elimination ordering for the components of the ODE system.

See also: tfpv_groebner_basis, print_results, print_tfpv, print_varieties

TikhonovFenichelReductions.tfpv_groebner_basis — Method

tfpv_groebner_basis(
    problem::ReductionProblem
) -> Vector{Nemo.QQMPolyRingElem}

Use the necessary conditions imposed by the RHS of the polynomial ODE System and its Jacobian to filter out TFPV candidates. All TFPVs must result in the vanishing of the returned set of polynomials.

Description

If π⁺ is a TFPV yielding a reduction onto an s-dimensional slow manifold, there exist a point x₀, such that

f(x₀,π⁺)=0

for any k>s the determinants of all k×k minors of D₁f(x₀,π⁺) vanish

These properties can be used to filter out possible TFPV candidates.

We are interested in partial solutions of the system of polynomials defined by the conditions above. In particular, we only consider conditions on the parameters, since there might be multiple slow manifolds and reductions. Thus, we eliminate the dynamic variables from the ideal and obtain a set of polynomials whose variables are the parameters of the ODE system. This function computes a generating set for this elimination ideal and all TFPVs lie in its vanishing set.

See also: tfpv_candidates

Computing a Tikhonov-Fenichel Reduction

TikhonovFenichelReductions.Reduction — Type

mutable struct Reduction

Type that holds all information to compute a Tikhonov-Fenichel reduction for a given slow-fast separation of rates.

Fields

idx::Vector{Bool}: slow-fast separation (0: slow, 1: fast)

s::Int: Dimension of reduced system (= dimension of slow manifold)

R::QQMPolyRing: Ring over rationals in x and p

x::Vector{QQMPolyRingElem}: Dynamic variables of system

p::Vector{QQMPolyRingElem}: All parameters

_p::Vector{QQMPolyRingElem}: All parameters, where slow parameters are set to 0

p_sf::Vector{QQMPolyRingElem}: Parameters, that are considered to be either small or large (all others are considered fixed)

idx_slow_fast::Vector{Bool}: Boolean indices, s.t. p_sf=p[idx_slow_fast]

f::Vector{QQMPolyRingElem}: RHS of system as vector with elements of ring R

f0::Vector{QQMPolyRingElem}: Fast / unperturbed part of system as vector with elements of ring R

f1::Vector{QQMPolyRingElem}: Slow / perturbed part of system as vector with elements of ring R

Df::AbstractAlgebra.Generic.MatSpaceElem{QQMPolyRingElem}: Jacobian of f

Df_x0::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Jacobian of f at non-singular point x0

T::AbstractAlgebra.Generic.PolyRing{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Ring in x over Fraction field K

chi::AbstractAlgebra.Generic.Poly{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Characteristic polynomial of Df_x0

M::Vector{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Slow manifold defined in all components of system

x0::Vector{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Non-singular point in the irreducible component of V(f0) containing the slow manifold

K::AbstractAlgebra.Generic.FracField{QQMPolyRingElem}: Fraction field in p

P::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Matrix with rational functions, such that f0=P⋅Psi

Psi::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Vector with polynomials, such that f0=P⋅Psi

DPsi::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Jacobian of Psi

success::Vector{Bool}: Indicates whether slow manifold M, non-singular point x0 and product decomposition f0=P⋅Psi have been set successfully

TikhonovFenichelReductions.Reduction — Method

Reduction(
    problem::ReductionProblem,
    idx::Vector{Bool};
    s
) -> Reduction

Constructor for Reduction Type.

Arguments

problem: Reduction problem type holding information on system and dimension of reduction.

idx: Boolean index indicating slow-fast separation of rates (0: small, 1: large).

s::Int: (optional) Dimension of slow manifold. Can be specified if a reduction corresponding to a TFPV for dimension different from problem.s should be considered (e.g. for manually computing a reduction for a given slow-fast separation that is not necessarily obtained via tfpv_candidates).

See also: set_manifold! set_decomposition!

TikhonovFenichelReductions.compute_reduction — Method

compute_reduction(
    reduction::Reduction
) -> Tuple{Union{Nothing, Vector{AbstractAlgebra.Generic.FracFieldElem{Nemo.QQMPolyRingElem}}}, Union{Nothing, Vector}}

Compute the reduced system after the slow manifold, non-singular point and product decomposition have been set successfully.

The function returns a tuple containing the reduced system in raw form and with variables substituted according to the slow manifold.

See also: set_manifold!, set_decomposition!, set_point!, Reduction

TikhonovFenichelReductions.jacobian_tfpv_at_x0 — Method

jacobian_tfpv_at_x0(
    reduction::Reduction
) -> AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{Nemo.QQMPolyRingElem}}

Return the Jacobian D₁f(x₀,π⁺) at the point x₀ on the slow manifold for a TFPV π⁺.

See also: Reduction, set_point!, set_manifold!

TikhonovFenichelReductions.jacobian_tfpv_on_manifold — Method

jacobian_tfpv_on_manifold(
    reduction::Reduction
) -> AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{Nemo.QQMPolyRingElem}}

Return the Jacobian D₁f(x,π⁺) at a generic point x on the slow manifold for a TFPV π⁺.

See also: Reduction, set_manifold!

TikhonovFenichelReductions.set_decomposition! — Method

set_decomposition!(
    reduction::Reduction,
    P::AbstractAlgebra.Generic.MatSpaceElem,
    Psi
) -> Bool

Set product decomposition f0=P⋅Psi locally satisfying 𝑉(f0)=𝑉(Psi), where P is a matrix of rational functions and Psi is a vector of polynomials.

Variants:

set_decomposition!(reduction::Reduction, P::AbstractAlgebra.Generic.MatSpaceElem, Psi): Manually specify P and Psi

set_decomposition!(reduction::Reduction, Psi): Try to compute P automatically. This works always for s=1, but may fail if s>1.

Description

Typically, Psi can be chosen from s independent entries of f0. Practically one can consider the generators of the ideals defining the irreducible components of 𝑉(f0) as entries for Psi (possibly rewriting the rational equations as polynomials by multiplying appropriately with parameters occurring in a denominator).

See also: set_manifold! set_point! Reduction

TikhonovFenichelReductions.set_manifold! — Method

set_manifold!(
    reduction::Reduction,
    M::AbstractVector
) -> Bool

Set the slow manifold by defining the values of the components of the system. Note that M must be defined as a vector with the same length as the system's components, i.e. reduction.x.

See also: Reduction, set_decomposition!, set_point!

TikhonovFenichelReductions.set_point! — Method

set_point!(reduction::Reduction, x0::AbstractVector) -> Bool

Set non-singular point on irreducible component of V(f⁰) corresponding to the slow manifold. Typically, this can be done automatically by setting the slow manifold.

See also: set_manifold!, set_decomposition!, Reduction

TikhonovFenichelReductions.show_slow_fast — Method

show_slow_fast(reduction::Reduction)

Print slow and fast parameters.

TikhonovFenichelReductions.slow_manifolds — Method

slow_manifolds(
    problem::ReductionProblem,
    idx::Vector{Bool}
) -> Tuple{Any, Any}

Compute the irreducible components of V(f⁰) and their dimensions for a given TFPV candidate. If there exist a reduction, the corresponding slow manifold must be contained in one of these components.

Arguments

problem: Reduction problem type holding information on system and dimension of reduction.

idx: Boolean index indicating slow-fast separation of rates (0: small, 1: large).

Description

This function can be used if one wants to check whether a particular slow-fast separation of rates yields a reduction for any dimension. If the dimension of an irreducible component of V(f⁰) differs from what was defined with ReductionProblem, the constructor Reduction can be called with the additional argument s specifying the dimension.

See also: Reduction

Output

TikhonovFenichelReductions.print_results — Function

print_results(
    problem::ReductionProblem,
    idx::Vector{Vector{Bool}},
    V,
    dim_V::Vector{Vector{Int64}};
    idx_subset
)

Print slow-fast separations that are TFPVs and generators for the corresponding irreducible components of V(f⁰) together with their Krull dimension.

Arguments

problem: ReductionProblem

idx: Boolean indices defining all TFPVs π⁺ that are slow-fast separations (0: slow, 1: fast).

V: Generators for the irreducible component of the affine varietiy V(f(⋅,π⁺)) for each slow-fast separation.

dim_V: Krull dimensions of the ideals generated by the elements in V (this equals the dimension of V(f(⋅,π⁺)) in the affine space ℂⁿ or its real part, given that it contains a real non-singular point).

idx_subset: (optional) Boolean index vector to include only certain TFPVs

See also: tfpv_candidates

TikhonovFenichelReductions.print_tfpv — Function

print_tfpv(
    problem::ReductionProblem,
    idx::Vector{Vector{Bool}};
    latex,
    idx_subset
)

Print TFPV (candidates) to terminal or return LaTeXString that can be included in tex file.

TikhonovFenichelReductions.print_varieties — Function

print_varieties(
    V::Array{Array{Array{T<:AbstractAlgebra.Generic.MPoly, 1}, 1}, 1};
    dim_V,
    latex,
    idx_subset
) -> Union{Nothing, LaTeXStrings.LaTeXString}

Print generators of ideals in corresponding to irreducible components of varieties for TFPV candidates to terminal (V and dim_V as returned by or tfpv_candidates).

See also: tfpv_candidates