API Reference

Finding Tikhonov-Fenichel Parameter Values

TikhonovFenichelReductions.ReductionProblem — Type

mutable struct ReductionProblem

Type that defines a Tikhonov-Fenichel reduction problem, i.e. a polynomial ODE system, for which all slow-fast separations of rates yielding a reduction onto dimension s are considered.

Fields

f::Vector{QQMPolyRingElem}: RHS of ODE system as a vector of polynomials

x::Vector{QQMPolyRingElem}: Vector of dynamic variables

θ::Vector{QQMPolyRingElem}: Vector of all parameters

s::Int: Dimension of reduced system

π::Vector{QQMPolyRingElem}: Vector of parameters to be considered slow or fast

idx_slow_fast::Vector{Bool}: Boolean index, such that π=ϑ[idx_slow_fast]

J::AbstractAlgebra.Generic.MatSpaceElem{QQMPolyRingElem}: Jacobian of f

_f::Function: RHS of ODE system as a Julia function with arguments x and ϑ

The type QQMPolyRingElem is used in Oscar.jl to represent elements of a polynomial ring; here this is ℝ[x,θ].

TikhonovFenichelReductions.ReductionProblem — Method

ReductionProblem(
    f::Function,
    x::Vector{String},
    θ::Vector{String},
    s::Int64;
    idx_slow_fast
) -> ReductionProblem

Constructor for ReductionProblem Type.

Arguments

f(x,θ)::Function: Julia function defining the RHS of system

x::Vector{String}: Vector of dynamic variables

θ::Vector{String}: Vector of all parameters

s::Int: Dimension of reduced system

idx_slow_fast::Vector{Bool}: Boolean index for all rates that are either small or large (all others are considered fixed)

Description

This function is used to setup the problem of finding Tikhonov-Fenichel Parameter Values, i.e. slow-fast separations of rates that yield a reduction onto an s-dimensional system. The names of all variables and parameters in the system are parsed to appropriate types in Oscar.jl, so that the necessary conditions for the existence of such a reduction can be evaluated.

See also: tfpv_candidates

TikhonovFenichelReductions.filter_determinants — Method

filter_determinants(
    problem::ReductionProblem;
    idx,
    multithreading
) -> Tuple{Vector{Vector{Bool}}, Vector{Nemo.QQMPolyRingElem}}

Use the conditions imposed by the RHS of the polynomial ODE System and its Jacobian to filter out TFPV candidates.

Description

If π⁺ is a TFPV yielding a reduction onto an s-dimensional slow manifold, there exist a point x₀, such that

f(x₀,π⁺)=0

for any k>s the determinants of all k×k minors of D₁f(x₀,π⁺) vanish

τ divides the characteristic polynomial Χ(τ) of D₁f(x₀,π⁺) exactly with power s

These properties can be used to filter out possible TFPV candidates.

By default, all 2ᵐ-2 possible slow-fast separations of the m parameters are considered, but if idx::Vector{Vector{Bool}} is defined, checking of the conditions is only performed on those candidates.

This function can be multithreaded by setting multithreading=true. Note that multithreading has to be enabled when the julia session is started (e.g. by julia --threads=auto, see the manual).

See also: filter_dimension

TikhonovFenichelReductions.filter_dimension — Method

filter_dimension(
    problem::ReductionProblem;
    idx,
    compute_primary_decomposition,
    exact_dimension
) -> Tuple{Vector{Vector{Bool}}, Tuple{Union{Nothing, Vector{Any}}, Union{Nothing, Vector{Vector{Int64}}}}}

Use the Krull dimension of the ideal generated by the unperturbed part of the system to filter out TFPV candidates form all possible slow-fast separations.

Description

The slow manifold on which the reduced system is defined is contained in 𝑉(f⁰), the affine variety of f⁰, i.e. the zero set of the fast / unperturbed part of the system (when we consider the entries as polynomials in the dynamic varieties x, so that the variety is a subset of the phase space). Thus, 𝑉(f⁰) needs to have an irreducible component with dimension s, which we can check using the Krull dimension of the corresponding ideal.

By default, all 2ᵐ-2 possible slow-fast separations of the m parameters are checked, but if idx::Vector{Vector{Bool}} is defined, only those candidates are considered (0: slow variable, 1: fast variable).

If compute_primary_decomposition=true (default behaviour) is set, the function attempts to compute a primary decomposition of the ideal corresponding to the unperturbed part of the system f⁰.

If in addition, exact_dimension=true (default), only those candidates are kept for which the affine variety has exactly dimension problem.s. Keep in mind that this is the Krull dimension, so the topological dimension over the reals can be smaller. However, the existence of a non-singular point in the irreducible component implies that topological and Krull dimension are equal. Since we require such a point later, we may use the exact dimension already.

See also: filter_determinants

TikhonovFenichelReductions.tfpv_candidates — Method

tfpv_candidates(
    problem;
    compute_primary_decomposition,
    exact_dimension,
    multithreading
)

Compute all TFPV candidates for the defined reduction problem by using the determinants criteria and asserting correct dimension of the slow manifold. This combines the functions filter_determinants and filter_dimension.

See also filter_determinants, filter_dimension

Computing a Tikhonov-Fenichel Reduction

TikhonovFenichelReductions.Reduction — Type

mutable struct Reduction

Type that holds all information to compute a Tikhonov-Fenichel reduction for a given slow-fast separation of rates.

Fields

idx::Vector{Bool}: slow-fast separation (0: slow, 1: fast)

s::Int: Dimension of reduced system (= dimension of slow manifold)

R::QQMPolyRing: Ring over rationals in x and θ

x::Vector{QQMPolyRingElem}: Dynamic variables of system

θ::Vector{QQMPolyRingElem}: All parameters

_θ::Vector{QQMPolyRingElem}: All parameters, where slow parameters are set to 0

π::Vector{QQMPolyRingElem}: Parameters, that are considered to be either small or large

idx_slow_fast::Vector{Bool}: Boolean indices, s.t. π=θ[idx_slow_fast]

f::Vector{QQMPolyRingElem}: RHS of system as vector with elements of ring R

f⁰::Vector{QQMPolyRingElem}: Fast / unperturbed part of system as vector with elements of ring R

f¹::Vector{QQMPolyRingElem}: Slow / perturbed part of system as vector with elements of ring R

Df::AbstractAlgebra.Generic.MatSpaceElem{QQMPolyRingElem}: Jacobian of f

Df_x₀::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Jacobian of f at non-singular point x₀

T::AbstractAlgebra.Generic.PolyRing{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Ring in x over Fraction field K

chi::AbstractAlgebra.Generic.Poly{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Characteristic polynomial of Df_x₀

M::Vector{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Slow manifold defined in all components of system

x₀::Vector{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Non-singular point in the irreducible component of V(f⁰) containing the slow manifold

K::AbstractAlgebra.Generic.FracField{QQMPolyRingElem}: Fraction field in θ

P::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Matrix with rational functions, such that f⁰=P⋅ψ

ψ::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Vector with polynomials, such that f⁰=P⋅ψ

Dψ::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.FracFieldElem{QQMPolyRingElem}}: Jacobian of ψ

success::Vector{Bool}: Indicates whether slow manifold M, non-singular point x₀ and product decomposition f⁰=P⋅ψ have been set successfully

TikhonovFenichelReductions.Reduction — Method

Reduction(
    problem::ReductionProblem,
    idx::Vector{Bool}
) -> Reduction

Constructor for Reduction Type.

Arguments

problem::ReductionProblem: Reduction problem type holding information on system and dimension of reduction.

idx::Union{Vector{Bool}, Vector{Int}}: Boolean index indicating slow-fast separation of rates (0: small, 1: large).

See also: set_manifold!, set_decomposition!

TikhonovFenichelReductions.compute_reduction — Method

compute_reduction(
    reduction::Reduction
) -> Tuple{Union{Nothing, Vector{AbstractAlgebra.Generic.FracFieldElem{Nemo.QQMPolyRingElem}}}, Union{Nothing, Vector{_A} where _A}}

Compute the reduced system after the slow manifold, non-singular point and product decomposition have been set successfully.

The function returns a tuple containing the reduced system in raw form and with variables substituted according to the slow manifold.

See also: set_manifold!, set_decomposition!, set_point!, Reduction

TikhonovFenichelReductions.set_decomposition! — Method

set_decomposition!(
    reduction::Reduction,
    P::AbstractAlgebra.Generic.MatSpaceElem,
    ψ
) -> Any

Set product decomposition f⁰=P⋅ψ locally satisfying 𝑉(f⁰)=𝑉(ψ), where P is a matrix of rational functions and ψ is a vector of polynomials.

Variants:

set_decomposition!(reduction::Reduction, P::AbstractAlgebra.Generic.MatSpaceElem, ψ): Manually specify P and ψ

set_decomposition!(reduction::Reduction, ψ): Try to compute P automatically. This works always for s=1, but may fail if s>1.

Description

Typically, ψ can be chosen from s independent entries of f⁰. Practically one can consider the generators of the ideals defining the irreducible components of 𝑉(f⁰) as entries for ψ (possibly rewriting the rational equations as polynomials by multiplying appropriately with parameters occurring in a denominator).

See also: set_manifold!, set_point!, Reduction

TikhonovFenichelReductions.set_manifold! — Method

set_manifold!(
    reduction::Reduction,
    M::AbstractVector{T} where T
) -> Bool

Set the slow manifold by defining the values of the components of the system. Note that M must be defined as a vector with the same length as the system's components, i.e. reduction.x.

See also: Reduction, set_decomposition!, set_point!

TikhonovFenichelReductions.set_point! — Method

set_point!(
    reduction::Reduction,
    x₀::AbstractVector{T} where T
) -> Bool

Set non-singular point on irreducible component of V(f⁰) corresponding to the slow manifold. Typically, this can be done automatically by setting the slow manifold.

See also: set_manifold!, set_decomposition!, Reduction

Output

TikhonovFenichelReductions.print_candidates — Function

print_candidates(
    idx::Vector{Vector{Bool}},
    prob::ReductionProblem;
    latex
) -> Union{Nothing, LaTeXStrings.LaTeXString}

Print TFPV candidates to terminal or return LaTeXString that can be included in tex file.

TikhonovFenichelReductions.print_varieties — Function

print_varieties(
    V,
    prob::ReductionProblem;
    latex
) -> Union{Nothing, LaTeXStrings.LaTeXString}

Print generators of ideals in corresponding to irreducible components of varieties for TFPV candidates to terminal (V as returned by filter_dimension or tfpv_candidates).

See also: filter_dimension, tfpv_candidates